Add to-yuv option
Add a new option --to-yuv (or -t) which converts each frame from rgb0/bgr0
to nv12/yuv420 in software using sws_scale(), before sending to the gpu
for encoding. This is needed for some drivers like the radeonsi vaapi driver,
which crashes when fed rgb0 data for encoding. See fdo bug #110719. Fixes #13.