Fix newlines in --help output
The newline characters within R"()" where printed literally and not used
as newlines. Fix this by adding printf specifically for this case. Also
add a newline to the end of the --help output so the command returns the
prompt after the help text and not at the end of the last line.