Fix: Crash when output is switched on
Currently, `chosen_output` is a pointer to an element in the `available_outputs`
vector. When an output is plugged in or switched on, this vector grows, is
reallocated and `chosen_output` becomes invalid, resulting in a segmentation fault.

Fixing this by using `std::list` instead of std::vector` for `available_outputs`.