Gracefully terminate on SIGTERM/SIGHUP to avoid corruption
If wf-recorder doesn't terminate gracefully, including closing the ffmpeg
stream, the generated recording is possibly corrupted, and can't be readily
played (with problems such as the infamous "moov atom not found" ffmpeg error)

wf-recorder _does_ gracefully terminate on a SIGINT signal, which is usually
generated by pressing Ctrl+C on the terminal. However, when wf-recorder was
terminated with a SIGTERM/SIGHUP, the default signal action was applied, which
terminated it without proper cleanup, possibly creating a corrupted recording.

Since SIGTERM/SIGHUP generally imply the program should gracefully terminate
(and SIGTERM is the typical way to kill a backgrounded process), handle them
like SIGINT, by closing the recording and exiting the main loop.